feat(lib-ref-inq): move logic to project-access
