+++ packages/project-access/package.json
@@ -34,6 +34,7 @@
     "dependencies": {
         "@sap-ux/i18n": "workspace:*",
         "@sap-ux/ui5-config": "workspace:*",
+        "fast-xml-parser": "4.2.7",
         "findit2": "2.2.3",
         "mem-fs": "2.1.0",
         "mem-fs-editor": "9.4.0",
