+++ packages/project-access/test/test-data/libs/sap.reuse.ex.test.lib.attachmentservice/src/sap/reuse/ex/test/lib/attachmentservice/attachment/components/fscomponent/Component.js
@@ -1,13 +1,13 @@
-sap.ui.define(['sap/se/mi/plm/lib/attachmentservice/attachment/Component'], function(AttachmentComponent) {
+sap.ui.define(['sap/se/mi/plm/lib/attachmentservice/attachment/Component'], function (AttachmentComponent) {
     'use strict';
 
     var Component = AttachmentComponent.extend(
-        'sap.se.mi.plm.lib.attachmentservice.attachment.components.fscomponent.Component',
+        'sap.reuse.ex.test.lib.attachmentservice.attachment.components.fscomponent.Component',
         {
             metadata: {
                 id: 'attachmentServiceComponent',
                 manifest: 'json',
-                library: 'sap.se.mi.plm.lib.attachmentservice',
+                library: 'sap.reuse.ex.test.lib.attachmentservice',
                 publicMethods: ['save', 'cancel', 'refresh', 'getApplicationState', 'getAttachmentCount']
             }
         }
