+++ packages/project-access/test/test-data/libs/sap.reuse.ex.test.lib.attachmentservice/src/sap/reuse/ex/test/lib/attachmentservice/attachment/manifest.json
@@ -1,0 +1,93 @@
+{
+    "_version": "1.1.0",
+    "sap.app": {
+        "_version": "1.1.0",
+        "id": "sap.reuse.ex.test.lib.attachmentservice.attachment",
+        "type": "library",
+        "i18n": "i18n/i18n.properties",
+        "embeddedBy": "../",
+        "title": "{{title}}",
+        "description": "{{description}}",
+        "ach": "PLM-FIO-DMS",
+        "dataSources": {
+            "CV_ATTACHMENT_SRV_Entities": {
+                "uri": "/sap/opu/odata/sap/CV_ATTACHMENT_SRV/"
+            }
+        },
+        "resources": "resources.json",
+        "offline": false
+    },
+    "sap.ui": {
+        "_version": "1.1.0",
+        "technology": "UI5",
+        "deviceTypes": {
+            "desktop": true,
+            "tablet": true,
+            "phone": true
+        },
+        "supportedThemes": ["sap_hcb", "sap_bluecrystal", "sap_belize"]
+    },
+    "sap.ui5": {
+        "_version": "1.1.0",
+        "rootView": {
+            "viewName": "sap.reuse.ex.test.lib.attachmentservice.attachment.view.Attachment",
+            "type": "XML"
+        },
+        "autoPrefixId": true,
+        "dependencies": {
+            "libs": {
+                "sap.ui.core": {},
+                "sap.m": {},
+                "sap.ui.table": {},
+                "sap.reuse.ex.test.lib.attachmentservice": {},
+                "sap.ui.comp": {}
+            },
+            "components": {
+                "sap.reuse.ex.test.lib.attachmentservice.attachment.components.fscomponent": {
+                    "lazy": true
+                },
+                "sap.reuse.ex.test.lib.attachmentservice.attachment.components.stcomponent": {
+                    "lazy": true
+                }
+            }
+        },
+        "componentUsages": {
+            "freestyleAttachmentReuseComponent": {
+                "name": "sap.reuse.ex.test.lib.attachmentservice.attachment.components.fscomponent",
+                "settings": {}
+            },
+            "smarttemplateAttachmentReuseComponent": {
+                "name": "sap.reuse.ex.test.lib.attachmentservice.attachment.components.stcomponent",
+                "settings": {}
+            }
+        },
+        "models": {
+            "i18n": {
+                "type": "sap.ui.model.resource.ResourceModel",
+                "settings": {
+                    "bundleName": "sap.reuse.ex.test.lib.attachmentservice.attachment.i18n.i18n"
+                }
+            }
+        },
+        "handleValidation": true,
+        "config": {},
+        "routing": {},
+        "contentDensities": {
+            "compact": false,
+            "cozy": false
+        }
+    },
+    "sap.platform.abap": {
+        "_version": "1.1.0",
+        "uri": "/sap/bc/bsp/sap/plm_ath_cres1/"
+    },
+    "sap.platform.hcp": {
+        "_version": "1.1.0",
+        "uri": ""
+    },
+    "sap.fiori": {
+        "_version": "1.1.0",
+        "registrationIds": ["F1243"],
+        "archeType": "reusecomponent"
+    }
+}
