+++ packages/ui5-library-reference-inquirer/package.json
@@ -32,7 +32,6 @@
     "dependencies": {
         "@sap-ux/inquirer-common": "workspace:*",
         "@sap-ux/project-access": "workspace:*",
-        "fast-xml-parser": "4.2.7",
         "i18next": "23.5.1"
     },
     "devDependencies": {
