+++ packages/ui5-library-reference-inquirer/src/prompts/constants.ts
